Rearrange some code.

+++ services/vanguard/sql/evedata.sql
@@ -236,7 +236,7 @@ CREATE TABLE `jitaPrice` (
 CREATE TABLE `jumps` (
   `toSolarSystemID` int(10) unsigned NOT NULL DEFAULT '0',
   `fromSolarSystemID` int(10) unsigned NOT NULL DEFAULT '0',
-  `jumps` tinyint(3) unsigned NOT NULL DEFAULT '0',
+  `jumps` smallint(5) unsigned NOT NULL DEFAULT '0',
   `secureJumps` smallint(5) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`toSolarSystemID`,`fromSolarSystemID`),
   KEY `fromSolarSystemID` (`fromSolarSystemID`)
